moves region check to utils
